Remove pbnh.util
