+++ run_tests.sh
@@ -12,4 +12,4 @@ pipenv run pytest --cov --cov-config "$init_config" --cov-branch --cov-fail-unde
 rm "$init_config"
 pipenv run pytest --cov pbnh.cli --cov-branch --cov-fail-under 100 --cov-report term-missing -v tests/test_cli.py
 pipenv run pytest --cov pbnh.db --cov-branch --cov-fail-under 100 --cov-report term-missing -v tests/test_db.py
-pipenv run pytest --cov pbnh --cov-branch --cov-fail-under 80 --cov-report term-missing
+pipenv run pytest --cov pbnh --cov-branch --cov-fail-under 90 --cov-report term-missing
